provisioning-scripts: in-field secure config
This scripts provides methods to import a GPG signed encrypted config
file. The target usecase is to allow in-field endusers to replace
device configuration in a secret and controlled way.

provisioning-scripts-usb adds a hotplug script to mount and look for
a .prov file in a usb-storage connected to the device. All required
kernel modules must be previously installed on the device. It looks
for <mac>.prov, <board_id>.prov and default.prov in all mountable
partitions, including subfolders.

provisioning-scripts-web adds a cgi at /cgi/provision where the user
upload a .prov file

provisioning-scripts-publisher is used to create a new config.prov file
from an existing config.tgz backup.

Both usb and web interface are disable on the second time the device
boots. The user needs to reset the device before provisioning a new
config.prov.

provisionsing-scripts-* must be installed in squashfs in order to
survive the factory reset.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>
(cherry picked from commit b3ac43781683844ebbacfa0bdb8b7baca76235a6)